refactor:rename server config file
